Refactor: auth separated into jwt, account, token, email packages
