+++ src/main/java/com/mudda/backend/email/EmailService.java
@@ -1,14 +1,14 @@
 /**
  * ---------------------------------------------------------------
  * Project : Mudda
- * File    : VerifyRequest
+ * File    : EmailService
  * Author  : Vikas Kumar
  * Created : 14-01-2026
  * ---------------------------------------------------------------
  */
-package com.mudda.backend.auth;
+package com.mudda.backend.email;
 
-public record VerifyRequest(
-        String email
-) {
+public interface EmailService {
+
+    void sendVerificationEmail(String email, String token);
 }
