+++ src/main/java/com/mudda/backend/email/TempEmailService.java
@@ -6,7 +6,7 @@
  * Created : 14-01-2026
  * ---------------------------------------------------------------
  */
-package com.mudda.backend.auth;
+package com.mudda.backend.email;
 
 import org.springframework.stereotype.Service;
