+++ src/main/java/com/mudda/backend/jwt/JwtService.java
@@ -6,7 +6,7 @@
  * Created : 12-01-2026
  * ---------------------------------------------------------------
  */
-package com.mudda.backend.auth;
+package com.mudda.backend.jwt;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
