+++ src/main/java/com/mudda/backend/token/refresh/RefreshToken.java
@@ -6,7 +6,7 @@
  * Created : 13-01-2026
  * ---------------------------------------------------------------
  */
-package com.mudda.backend.auth;
+package com.mudda.backend.token.refresh;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
