+++ src/main/java/com/mudda/backend/token/verification/VerificationTokenRepository.java
@@ -6,8 +6,9 @@
  * Created : 14-01-2026
  * ---------------------------------------------------------------
  */
-package com.mudda.backend.auth;
+package com.mudda.backend.token.verification;
 
+import com.mudda.backend.token.TokenType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
